+++ android/festago/domain/src/main/java/com/festago/festago/repository/StudentsVerificationRepository.kt
@@ -1,5 +1,5 @@
 package com.festago.festago.repository
 
 interface StudentsVerificationRepository {
-    suspend fun sendVerificationCode(userName: String, schoolId: Int): Result<Unit>
+    suspend fun sendVerificationCode(userName: String, schoolId: Long): Result<Unit>
 }
